fix(mobile): fix mobile components build error
