+++ packages/mobile/components/multi-select-item/src/mobile.vue
@@ -49,10 +49,10 @@
 </template>
 
 <script lang="ts">
-import { $prefix, props, setup, defineComponent } from '../../../vue-common'
+import { $prefix, setup, defineComponent } from '../../../vue-common'
 import { renderless, api } from './renderless/vue'
 import { iconChevronDown } from '@opentiny/vue-icon'
-import Button from '@opentiny/vue-button'
+import Button from '../../button'
 import { multiSelectItemProps } from './multi-select-item'
 import '@opentiny/vue-theme-mobile/multi-select-item/index.less'
