+++ packages/mobile/components/upload/src/mobile.vue
@@ -11,7 +11,7 @@
  -->
 
 <script lang="tsx">
-import { $prefix, setup, h, defineComponent, props } from '@opentiny/vue-common'
+import { $prefix, setup, h, defineComponent } from '../../../vue-common'
 
 // 此处引入 h 是为了防止打包后 h 被重命名导致组件报错的问题
 import { renderless, api } from './renderless/vue'
@@ -37,7 +37,6 @@ export default defineComponent({
     let {
       accept,
       disabled,
-      drag,
       handleChange,
       handleClick,
       handlePaste,
@@ -46,7 +45,6 @@ export default defineComponent({
       listType,
       multiple,
       name,
-      uploadFiles,
       fileList,
       limit,
       isHidden,
