+++ packages/mobile/vite.config.ts
@@ -1,19 +1,16 @@
 import { defineConfig } from 'vite'
 import vue from '@vitejs/plugin-vue'
 import vueJsx from '@vitejs/plugin-vue-jsx'
-import dts from 'vite-plugin-dts'
 import svgLoader from 'vite-svg-loader'
 
-// https://vitejs.dev/config/
 export default defineConfig({
-  plugins: [vue(), vueJsx(), dts(), svgLoader()],
+  plugins: [vue(), vueJsx(), svgLoader()],
   build: {
     lib: {
       entry: './index.ts'
     },
-    minify: true,
     rollupOptions: {
-      external: [/@better-scroll/, /@opentiny/],
+      external: [/@opentiny\/vue/, /@better-scroll/, 'vue', 'xss'],
       input: ['index.ts'],
       output: [
         {
